category confirmation dialog, error handling in form, form reset

+++ insertApp/src/app/pages/category-view/category-view.component.ts
@@ -5,6 +5,9 @@ import {Category, Subcategory} from 'src/app/models/category';
 import { FileUploadService } from 'src/app/file-upload.service';
 import { map, share } from 'rxjs/operators';
 import { DomSanitizer } from '@angular/platform-browser';
+import {MatSnackBar} from '@angular/material/snack-bar'
+import { ViewChild } from '@angular/core';
+
 
 
 
@@ -15,30 +18,48 @@ import { DomSanitizer } from '@angular/platform-browser';
 })
 export class CategoryViewComponent implements OnInit{
 
+  @ViewChild('imageUpload') inputCat: ElementRef;
+  @ViewChild('imageUploadSubcat') inputSubcat: ElementRef;
+  @ViewChild('imageUploadUpdateSubcategory') inputUpdateSubcat: ElementRef;
+
+  imageUploadUpdateSubcategory
+
   categoryName = new FormControl('')
   subcategoryName = new FormControl('')
   categories: Category[]
   image: File
   chosen:boolean
   uploadedFile:any
-
   clickedSubcategory: number;
-
-
+  private _allowed_image_types = ["image/png", "image/jpeg", "image/jpg"]
+  private _max_image_size = 10000
   constructor(
     private categoryService: CategoryService,
     private fileService: FileUploadService,
-    private sanitizer: DomSanitizer
+    private sanitizer: DomSanitizer,
+    private _snackbar: MatSnackBar,
   ) { }
+
 
   ngOnInit(): void {
     this.categoryService.categories.subscribe(cat => {
       this.categories = cat
     });
   }
+
   addNewCategory():void {
     let category = new Category();
     category.name = this.categoryName.value;
+    if(this.image && this.image.size > this._max_image_size){
+      this.openSnackBar('File size too big!', 'error')
+      this.resetForms()
+      return
+    }
+    if(this.image && !this._allowed_image_types.includes(this.image.type)){
+      this.openSnackBar('File type ' + this.image.type + ' is not allowed', 'error')
+      this.resetForms()
+      return
+    }
     if(this.image){
       this.categoryService.createCategory(category).pipe(
         map(catRes => {
@@ -49,28 +70,46 @@ export class CategoryViewComponent implements OnInit{
             formdata.append('file_path', categoryImagePath)
             this.fileService.uploadFile(formdata).subscribe((response)=> {
               category.iconPath = response.path
-              this.categoryService.updateCategory(category._id, category).subscribe(x => console.log('cat: ', x) )
+              this.categoryService.updateCategory(category._id, category).subscribe(
+                  category => this.openSnackBar("Successfully uploaded category: "+ category.name, 'success'),
+                  err => this.openSnackBar("An error occured: "+ err, 'error'),
+              )
           })
         }),
         share()
-        ).toPromise().then(() => { console.log(this.categories)
-    })
+        ).toPromise().then(() => {
+        this.resetForms()
+      })
       }
       else{
-          console.error('No image uploaded, but its necessary for a category!')
-        }
+        this.openSnackBar('No image uploaded, but its necessary for a category!', 'error')
+        this.resetForms()
+
+    }
   }
 
   deleteCategory(category: Category): void {
-    this.categoryService.deleteCategory(category._id).subscribe()
+    if(confirm("Are you sure to delete Category "+ category.name +
+        "\nYou are not able to find the related events again")) {
+      this.categoryService.deleteCategory(category._id).subscribe()
+    }
   }
 
   addNewSubcategory(category: Category): void {
-    let subcategory = new Subcategory()
-    subcategory.name = this.subcategoryName.value
-    category.subcategories.push(subcategory)
-
+    if(this.image.size > this._max_image_size){
+      this.openSnackBar('File size too big!', 'error')
+      this.resetForms()
+      return
+    }
+    if(!this._allowed_image_types.includes(this.image.type)) {
+      this.openSnackBar('File type ' + this.image.type + ' is not allowed', 'error')
+      this.resetForms()
+      return
+    }
     if(this.image){
+      let subcategory = new Subcategory()
+      subcategory.name = this.subcategoryName.value
+      category.subcategories.push(subcategory)
       this.categoryService.updateCategory(category._id, category).subscribe(
           catRes => {
             category = catRes
@@ -85,20 +124,26 @@ export class CategoryViewComponent implements OnInit{
               category.subcategories.map(sub => {
                 if(sub._id == subcategory._id){ sub.iconPath = response.path }
               })
-              this.categoryService.updateCategory(category._id, category).subscribe(x => console.log('cat: ', x) )
+              this.categoryService.updateCategory(category._id, category).subscribe(
+                  category => this.openSnackBar("Successfully uploaded subcategory: "+ subcategory.name, 'success'),
+                  err => this.openSnackBar("An error occured: "+ err, 'error'),
+                  () => this.resetForms())
             })
           },
       )
     }
     else{
-      console.error('No image uploaded, but its necessary for a category!')
+      this.openSnackBar("No image icon, but its necessary for a subcategory!", 'error')
     }
 
   }
 
   deleteSubcategory(category: Category, subcategory): void {
-    category.subcategories = category.subcategories.filter(subcat => subcat.name !== subcategory.name )
-    this.categoryService.updateCategory(category._id, category).subscribe()
+    if(confirm("Are you sure to delete Subcategory "+ subcategory.name +
+        "\nYou are not able to find the related events again")) {
+      category.subcategories = category.subcategories.filter(subcat => subcat.name !== subcategory.name)
+      this.categoryService.updateCategory(category._id, category).subscribe()
+    }
   }
   fileChosen(event: any){
     if(event.target.value){
@@ -125,9 +170,7 @@ export class CategoryViewComponent implements OnInit{
     )
     return iconURL
   }
-  uploadNewSubcategoryIcon(){
-  return 0
-  }
+
   clickSubcategory(id){
     console.log(id, this.clickedSubcategory)
     if(id == this.clickedSubcategory) {
@@ -136,6 +179,29 @@ export class CategoryViewComponent implements OnInit{
     else{
       this.clickedSubcategory=id
     }
+  }
+
+  updateSubcategoryIcon(){
+    console.log("to implement")
+    this.resetForms()
+  }
+
+  openSnackBar(message, state){
+
+    this._snackbar.open(message, '' , {
+      duration: 3000,
+      verticalPosition: 'top',
+      horizontalPosition: 'center',
+      panelClass: [state !== 'error' ? 'green-snackbar' : 'red-snackbar']
+
+    });
+  }
+  resetForms(){
+    this.categoryName.reset()
+    this.subcategoryName.reset()
+    this.inputCat.nativeElement.value = '';
+    this.inputSubcat.nativeElement.value = '';
+    this.inputUpdateSubcat.nativeElement.value = '';
   }
 
 }
